vyos.utils: T7648: only provide one chown function
Remove duplicate function definition vyos.utils.file.chown which was also
provided by vyos.utils.permission.chown.
