+++ python/vyos/utils/file.py
@@ -92,36 +92,6 @@ def read_json(fname, defaultonfailure=None):
             return defaultonfailure
         raise e
 
-def chown(path, user=None, group=None, recursive=False):
-    """ change file/directory owner """
-    from pwd import getpwnam
-    from grp import getgrnam
-
-    if user is None and group is None:
-        return False
-
-    # path may also be an open file descriptor
-    if not isinstance(path, int) and not os.path.exists(path):
-        return False
-
-    # keep current value if not specified otherwise
-    uid = -1
-    gid = -1
-
-    if user:
-        uid = getpwnam(user).pw_uid
-    if group:
-        gid = getgrnam(group).gr_gid
-
-    if recursive:
-        for dirpath, dirnames, filenames in os.walk(path):
-            os.chown(dirpath, uid, gid)
-            for filename in filenames:
-                os.chown(os.path.join(dirpath, filename), uid, gid)
-    else:
-        os.chown(path, uid, gid)
-    return True
-
 
 def chmod(path, bitmask):
     # path may also be an open file descriptor
