+++ src/conf_mode/service_dhcpv6-server.py
@@ -25,11 +25,11 @@ from vyos.config import Config
 from vyos.template import render
 from vyos.utils.process import call
 from vyos.utils.file import chmod_775
-from vyos.utils.file import chown
 from vyos.utils.file import makedir
 from vyos.utils.file import write_file
 from vyos.utils.dict import dict_search
 from vyos.utils.network import is_subnet_connected
+from vyos.utils.permission import chown
 from vyos import ConfigError
 from vyos import airbag
 airbag.enable()
